requirements.txt
connexion
cryptography
pyjwt
python-dotenv
pytz
six
swagger-ui-bundle # connexion[swagger-ui]
uwsgi
